refactor return_menu
