+++ menus/return_menu/return_answers.py
@@ -1,0 +1,7 @@
+def back():
+    return
+
+
+return_aswers = {
+    'Назад': back
+}
