+++ menus/return_menu/return_questions.py
@@ -1,0 +1,2 @@
+return_questions = [{'type': 'list', 'name': 'return_menu', 'message': '', 'choices': [
+    'Назад']}]
