+++ validators/RequiredValidator.py
@@ -1,0 +1,7 @@
+from PyInquirer import Validator, ValidationError
+class RequiredValidator(Validator):
+    def validate(self, document):
+        if not len(document.text):
+            raise ValidationError(
+                message='Пожалуйста введите название',
+                cursor_position=len(document.text))
